feat: check owner before update & delete

+++ apps/api/src/handlers/edm.ts
@@ -69,6 +69,20 @@ edm.openapi(updateFile, async (c) => {
   await checkRole(roles, false);
   const { file, name, description } = c.req.valid('form');
 
+  const { data: isOwner, error: ownerError } = await supabase
+    .from('DOCUMENTS')
+    .select('owner')
+    .eq('name', name)
+    .single();
+
+  if (ownerError) {
+    return c.json({ error: ownerError.message }, 500);
+  }
+
+  if (isOwner && isOwner.owner !== user.id) {
+    return c.json({ error: 'You are not the owner of this document' }, 400);
+  }
+
   const { error: updateDoc } = await supabase.from('DOCUMENTS').update({ description }).eq('name', name);
   if (updateDoc) {
     return c.json({ error: 'Error updating document' }, 500);
@@ -87,6 +101,20 @@ edm.openapi(deleteFileRoute, async (c) => {
   const roles = user.roles;
   await checkRole(roles, false);
   const { name } = c.req.valid('json');
+
+  const { data: isOwner, error: ownerError } = await supabase
+    .from('DOCUMENTS')
+    .select('owner')
+    .eq('name', name)
+    .single();
+
+  if (ownerError) {
+    return c.json({ error: ownerError.message }, 500);
+  }
+
+  if (isOwner && isOwner.owner !== user.id) {
+    return c.json({ error: 'You are not the owner of this document' }, 400);
+  }
 
   const { error: deleteDoc } = await supabase.from('DOCUMENTS').delete().eq('name', name);
   if (deleteDoc) {
